Make user delete api test script 1. add user delete api test script [prototype] 2. function delete user could not be delete admin user

+++ app/server/module/manager/user_manager.py
@@ -324,6 +324,11 @@ class UserManager(WorkerManager):
         if req_static_id == target_static_id:
             raise MicrocloudchipAuthAccessError("User can't remove itself")
 
+        # Admin 계정 삭제 불가
+        target_is_admin = len(model.User.objects.filter(is_admin=True).filter(static_id=req_static_id))
+        if target_is_admin:
+            raise MicrocloudchipAuthAccessError("Admin could not be deleted")
+
         # Storage 제거
         try:
             storage_manager.delete_directory(target_static_id, {
